+++ spinetoolbox/project_items/tool/widgets/add_tool_widget.py
@@ -52,7 +52,7 @@ class AddToolWidget(QWidget):
         self.name = ''
         self.description = ''
         # Init
-        self.ui.comboBox_tool.setModel(self._toolbox.tool_template_model)
+        self.ui.comboBox_tool.setModel(self._toolbox.tool_specification_model)
         self.ui.lineEdit_name.setFocus()
         self.connect_signals()
         # Ensure this window gets garbage-collected when closed
@@ -67,7 +67,7 @@ class AddToolWidget(QWidget):
 
     @Slot(int, name='update_args')
     def update_args(self, row):
-        """Show Tool template command line arguments in text input.
+        """Show Tool specification command line arguments in text input.
 
         Args:
             row (int): Selected row number
@@ -76,7 +76,7 @@ class AddToolWidget(QWidget):
             # No Tool selected
             self.ui.lineEdit_tool_args.setText("")
             return
-        selected_tool = self._toolbox.tool_template_model.tool_template(row)
+        selected_tool = self._toolbox.tool_specification_model.tool_specification(row)
         args = selected_tool.cmdline_args
         if not args:
             # Tool cmdline_args is None if the line does not exist in Tool definition file